Refactor - 1st iteration
