+++ src/components/gtt-client/index.ts
@@ -1,0 +1,4 @@
+export * from './redmine';
+
+import GttClient from './gtt-client-class';
+export { GttClient };
