+++ src/components/gtt-client/interfaces.ts
