+++ src/index.ts
@@ -4,12 +4,12 @@ import './stylesheets/app.scss';
 
 // Custom Icons
 import './stylesheets/custom-icons.css';
-import './stylesheets/CustomIconsDef.js';
+import './stylesheets/custom-icons-def.js';
 
 // Material Design Icons
 // https://github.com/marella/material-design-icons/tree/main/font#readme
 import '@material-design-icons/font/filled.css';
-import './stylesheets/MaterialDesignDef.js';
+import './stylesheets/material-design-def.js';
 
 import { GttClient } from './components/gtt-client';
 import { gtt_setting } from './components/gtt-setting';
